+++ apps/web/app/(ee)/partners.dub.co/(apply)/[programSlug]/page.tsx
@@ -1,6 +1,6 @@
 import { getProgram } from "@/lib/fetchers/get-program";
 import { programLanderSchema } from "@/lib/zod/schemas/program-lander";
-import { BLOCK_COMPONENTS } from "@/ui/partners/lander-blocks";
+import { BLOCK_COMPONENTS } from "@/ui/partners/lander/blocks";
 import { LanderHero } from "@/ui/partners/lander/lander-hero";
 import { LanderRewards } from "@/ui/partners/lander/lander-rewards";
 import { notFound } from "next/navigation";
