+++ apps/web/ui/partners/lander/blocks/AccordionBlock.tsx
