+++ apps/web/ui/partners/lander/blocks/BlockDescription.tsx
