+++ apps/web/ui/partners/lander/blocks/ImageBlock.tsx
