+++ apps/web/ui/partners/lander/blocks/TextBlock.tsx
