+++ apps/web/ui/partners/lander/blocks/earnings-calculator-block.tsx
@@ -6,7 +6,7 @@ import { InvoiceDollar } from "@dub/ui";
 import NumberFlow from "@number-flow/react";
 import { useId, useState } from "react";
 import { z } from "zod";
-import { formatRewardDescription } from "../format-reward-description";
+import { formatRewardDescription } from "../../format-reward-description";
 import { BlockDescription } from "./BlockDescription";
 import { BlockTitle } from "./BlockTitle";
